rebar.config.script: replace NO_SYSTEM_RE2
While I would prefer to detect and use a system provided re2 lib by
default, this is likely to distrupt existing users' environments due to
the runtime requirement of libre2.so. Therefore, replace the
NO_SYSTEM_RE2 env var with SYSTEM_RE2. This way the existing build
behavior stays the same and distro packagers are given an option to link
against libre2.so found on the system.