rebar.config.script: print debug info to stderr
In order for rebar3 to not filter the config script's debug log, we need
to use standard_error. Otherwise, one would have to run the following:

$ DEBUG=1 DEBUG_CONFIG=1 rebar3 compile

This has the problem that we get rebar3's debug logs as well, but that's
not necessarily what we want.

With this commit, we can just run

$ DEBUG_CONFIG=1 rebar3 compile

and get normal logs plus the config script's debug log.