rebar.config.script: implement Elixir mix kludge
When a rebar project is a dependency of an Elixir project and it has a
rebar.config.script file, Elixir's mix tool will use its own not quite
correct evaluator to retrieve dependencies of such a project before
actually running rebar on it. To account for this situation, we adjust
the check for deciding if rebar's port compiler plugin is new enough to
assume mix and therefore by default rebar3 is used if we don't find the
app key rebar->vsn.